SerialStream works now but...
Okay, I got the serial stream to work and I have adapted the BufferedSerial to a stream instance in the constructor of the BNO8x library.

Now the issue I am having is that, whenever I compile the code, I get the following error mostly:
1. Use of undeclared identifier  BNO080Async

wait....what?!

So I realized that the BNO08x library uses RTOS and not baremetal, therefore I removed the json file that contained the baremetal profile.

Then I recompiled the flipping thing and it worked but....yes, there's always a but, but...nothing is been printed to the terminal.

Now what I will do is, I will use the example code provided for the IMU and adapt the serial communication appropriately and see what happens