Digital Position Feedback Implementation
Digital Feedback implemented.
Pressing 'e' extends the actuator
Pressing 'q' retracts the actuator

But the distance moved for either extension or retraction is know ONLY when 's' is pressed to stop.

Next goal would be to update the position of the stroke  without having to stop the actuator.

Whenever direction of movement is changed i.e. from extension to retraction, print out the previous position before direction change.

This would allow for position update whenever the actuator moves in the other direction without having to stop.