DPF Version 2 - implemented into a Class
This version of the DPF has been implemented into a class.

This allows for feedback readings for all six actuators.

The constructor of the class allows users to specify the RPWM and LPWM pins for each actuator as well as the speed of individual actuators and their duty cycle

It also uses the buffered serial UART class to print out the current position of the actuators.

Finally it uses ANSI characters to reset the terminal and set the cursor back to the home position (top left corner). This gives a much cleaner visual representation of the data been printed to the terminal